+++ packages/headless/src/controllers/knowledge/generated-answer/headless-generated-answer-with-follow-ups.ts
@@ -42,21 +42,41 @@ export interface GeneratedAnswerWithFollowUps extends GeneratedAnswer {
   state: GeneratedAnswerWithFollowUpsState;
   /**
    * Marks the answer as liked.
-   * @param answerId - Optional ID of the answer to like. Defaults to the head answer.
+   * @param answerId - Optional ID of the answer to like. Defaults to the first answer.
    */
   like(answerId?: string): void;
 
   /**
    * Marks the answer as disliked.
-   * @param answerId - Optional ID of the answer to dislike. Defaults to the head answer.
+   * @param answerId - Optional ID of the answer to dislike. Defaults to the first answer.
    */
   dislike(answerId?: string): void;
 
   /**
+   * Logs a custom event indicating a cited source link was hovered.
+   * @param citationId - The ID of the hovered citation.
+   * @param citationHoverTimeMs - The number of milliseconds spent hovering over the citation.
+   * @param answerId - Optional ID of the answer for which the citation was hovered. Defaults to the first answer.
+   */
+  logCitationHover(
+    citationId: string,
+    citationHoverTimeMs: number,
+    answerId?: string
+  ): void;
+
+  /**
+   * Logs a click on a cited source link for analytics.
+   * @param citationId - The ID of the clicked citation.
+   * @param answerId - Optional ID of the answer for which the citation was clicked. Defaults to the first answer.
+   */
+  logCitationClick(citationId: string, answerId?: string): void;
+
+  /**
    * Logs a copy-to-clipboard interaction for analytics.
-   * @param answerId - Optional ID of the copied answer. Defaults to the current answer.
+   * @param answerId - Optional ID of the copied answer. Defaults to the first answer.
    */
   logCopyToClipboard(answerId?: string): void;
+
   /**
    * Asks a follow-up question.
    * @param question - The follow-up question to ask.
@@ -120,6 +140,7 @@ export function buildGeneratedAnswerWithFollowUps(
         followUpAnswers: followUpAnswersState,
       };
     },
+
     retry() {
       answerRunner.run(
         engine.state,
@@ -127,28 +148,69 @@ export function buildGeneratedAnswerWithFollowUps(
         () => engine.navigatorContext
       );
     },
+
+    // TODO: SFINT-6665
     like(answerId?: string) {
       if (!answerId || this.state.answerId === answerId) {
         controller.like();
         return;
       }
-      engine.dispatch(likeFollowUp({answerId}));
+
+      if (!this.state.liked) {
+        engine.dispatch(likeFollowUp({answerId}));
+        engine.dispatch(analyticsClient.logLikeGeneratedAnswer(answerId));
+      }
     },
+
+    // TODO: SFINT-6665
     dislike(answerId?: string) {
       if (!answerId || this.state.answerId === answerId) {
         controller.dislike();
         return;
       }
-      engine.dispatch(dislikeFollowUp({answerId}));
+
+      if (!this.state.disliked) {
+        engine.dispatch(dislikeFollowUp({answerId}));
+        engine.dispatch(analyticsClient.logDislikeGeneratedAnswer(answerId));
+      }
     },
+
+    // TODO: SFINT-6665
     logCopyToClipboard(answerId?: string) {
       if (!answerId || this.state.answerId === answerId) {
         controller.logCopyToClipboard();
         return;
       }
-      // Todo: SFINT-6581 implement logCopyFollowUp action and dispatch here
-      console.warn(
-        'Method not yet implemented to send analytics for copy to clipboard on a followup answer'
+      engine.dispatch(analyticsClient.logCopyGeneratedAnswer(answerId));
+    },
+
+    // TODO: SFINT-6665
+    logCitationClick(citationId: string, answerId?: string) {
+      if (!answerId || this.state.answerId === answerId) {
+        controller.logCitationClick(citationId);
+        return;
+      }
+      engine.dispatch(
+        analyticsClient.logOpenGeneratedAnswerSource(citationId, answerId)
+      );
+    },
+
+    // TODO: SFINT-6665
+    logCitationHover(
+      citationId: string,
+      citationHoverTimeMs: number,
+      answerId?: string
+    ) {
+      if (!answerId || this.state.answerId === answerId) {
+        controller.logCitationHover(citationId, citationHoverTimeMs);
+        return;
+      }
+      engine.dispatch(
+        analyticsClient.logHoverCitation(
+          citationId,
+          citationHoverTimeMs,
+          answerId
+        )
       );
     },